feat: click event trigger setting

+++ lib/local/DwellClick/include/Click.h
@@ -14,7 +14,7 @@ namespace DwellClick {
     private:
         std::chrono::steady_clock::time_point fixation_start;  // 응시 시작 시간
         bool isCounting = false;  // 응시 시간 측정 중인지 여부
-        double fixation_threshold = 2.0;  // 응시 임계 시간 (초)
+        int fixation_threshold = 2;  // 응시 임계 시간 (초)
         double fixation_distance_threshold = 30.0;  // 고정 여부를 판단하는 거리 임계값
 
         // 클릭 이벤트 발생
